Update net_arp.c
I could be wrong but it would appear that the timeout function should be NetARP_CacheReqTimeout and not NetARP_CacheRenewTimeout for the arp renew.  Otherwise if there is no ARP reply, the NetARP_CacheReqTimeout function is never called and that's what frees the ARP from the table if there is no response.